repo/CognitoSync-NovachainNexus/requirements.txt
fastapi
uvicorn
python-dotenv
crewai
masumi
pydantic
python-multipart
httpx
langchain
langchain-community
